Version 1
This is the final script used in the Trip advisor project.  It has 4 packages: rvest, stringr, dplyr, tm.  It has 3 for loops form the original data set.  The data set is not added due to the privacy of the client.  But the code is able to share.

The final data set was then applied to google sheets for the answers.